pyproject.toml
[project]
name = "fifa-26"
version = "0.1.0"
description = "FIFA 26 REST API"
authors = [
    {name = "Praagnya"}
]
requires-python = ">=3.12"

dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "httpx>=0.28",
    "python-dotenv>=1.0",
    "google-genai>=1.0",
    "tavily-python>=0.5",
]

[project.urls]
Repository = "https://github.com/Praagnya/fifa-26"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "ruff>=0.9",
]